test-requirements.txt
asgiref==3.6.0
    # via
    #   -r requirements.txt
    #   django
build==0.9.0
    # via
    #   -r requirements.txt
    #   pip-tools
certifi==2022.12.7
    # via
    #   -r requirements.txt
    #   requests
charset-normalizer==2.1.1
    # via
    #   -r requirements.txt
    #   requests
click==8.1.3
    # via
    #   -r requirements.txt
    #   pip-tools
django==3.2.16
    # via
    #   -r requirements.txt
    #   django-more-admin-filters
    #   django-sslserver
django-more-admin-filters==1.3
    # via -r requirements.txt
django-nested-admin==4.0.2
    # via -r requirements.txt
django-querycount==0.7.0
    # via -r requirements.txt
django-sslserver==0.22
    # via -r requirements.txt
django-test-plus==2.2.1
    # via -r test-requirements.in
factory-boy==3.2.1
    # via -r test-requirements.in
faker==15.3.4
    # via factory-boy
idna==3.4
    # via
    #   -r requirements.txt
    #   requests
packaging==22.0
    # via
    #   -r requirements.txt
    #   build
pep517==0.13.0
    # via
    #   -r requirements.txt
    #   build
pip-tools==6.12.1
    # via -r requirements.txt
python-dateutil==2.8.2
    # via faker
python-monkey-business==1.0.0
    # via
    #   -r requirements.txt
    #   django-nested-admin
pytz==2022.7
    # via
    #   -r requirements.txt
    #   django
requests==2.28.1
    # via -r requirements.txt
six==1.16.0
    # via
    #   -r requirements.txt
    #   python-dateutil
    #   python-monkey-business
sqlparse==0.4.3
    # via
    #   -r requirements.txt
    #   django
tomli==2.0.1
    # via
    #   -r requirements.txt
    #   build
    #   pep517
urllib3==1.26.13
    # via
    #   -r requirements.txt
    #   requests
wheel==0.38.4
    # via
    #   -r requirements.txt
    #   pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
